+++ package.json
@@ -30,7 +30,8 @@
     "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "~3.4.0",
     "react-native-tab-view": "^3.1.1",
-    "react-native-pager-view": "^5.4.24"
+    "react-native-pager-view": "^5.4.24",
+    "@react-native-picker/picker": "^2.6.1"
   },
   "devDependencies": {
     "@babel/core": "^7.18.5",
